+++ src/utils/VariableProperty.test.ts
@@ -1,4 +1,4 @@
-import { describe, expect, it } from 'vitest';
+import { describe, expect, it, vi } from 'vitest';
 import { Validator, VariableProperty } from '.';
 
 describe('VariableProperty', () => {
@@ -43,6 +43,50 @@ describe('VariableProperty', () => {
 
         prop.current.foo = 'qux';
         expect(prop.invalidated).toBe(true);
+      });
+    });
+  });
+
+  describe('listeners', () => {
+    describe('when a scalar property changes', () => {
+      it('calls the listener function', () => {
+        const property = new VariableProperty(42);
+        const fn = vi.fn();
+
+        property.listen(fn);
+        expect(fn).not.toBeCalled();
+  
+        property.current = 43;
+        expect(fn).toBeCalledTimes(1);
+        expect(fn).toBeCalledWith(43);
+      });
+    });
+
+    describe('when an array property changes', () => {
+      it('calls the listener function', () => {
+        const property = new VariableProperty(['a', 'b']);
+        const fn = vi.fn();
+
+        property.listen(fn);
+        expect(fn).not.toBeCalled();
+  
+        property.current = ['c', 'd'];
+        expect(fn).toBeCalledTimes(1);
+        expect(fn).toBeCalledWith(['c', 'd']);
+      });
+    });
+
+    describe('when an entry of an array property changes', () => {
+      it('calls the listener function', () => {
+        const property = new VariableProperty(['a', 'b']);
+        const fn = vi.fn();
+
+        property.listen(fn);
+        expect(fn).not.toBeCalled();
+  
+        property.current[0] = 'c';
+        expect(fn).toBeCalledTimes(1);
+        expect(fn).toBeCalledWith(['c', 'b']);
       });
     });
   });